+++ hedera-mirror-rest/__tests__/specs/contracts/results/{id}/missing-transaction.json
@@ -49,7 +49,7 @@
     "_status": {
       "messages": [
         {
-          "message": "No correlating transaction"
+          "message": "Not found"
         }
       ]
     }
