+++ hedera-mirror-rest/__tests__/specs/contracts/{id}/results/{id}/non-populated-eth-data.json
@@ -33,7 +33,10 @@
         "gas_limit": 987654,
         "gas_used": 123,
         "payer_account_id": 8001,
-        "sender_id": 8001
+        "sender_id": 8001,
+        "transaction_hash": "4a563af33c4871b51a8b108aa2fe1dd5280a30dfb7236170ae5e5e7957eb6392",
+        "transaction_result": 22,
+        "transaction_index": 1
       }
     ],
     "contractlogs": [
