+++ hedera-mirror-rest/__tests__/specs/contracts/{id}/results/{id}/timestamp-result-not-found.json
@@ -14,7 +14,10 @@
         "function_result": [8, 8],
         "gas_limit": 987654,
         "gas_used": 123,
-        "payer_account_id": 8001
+        "payer_account_id": 8001,
+        "transaction_hash": "96ecf2e0cf1c8f7e2294ec731b2ad1aff95d9736f4ba15b5bbace1ad2766cc1c",
+        "transaction_result": 22,
+        "transaction_index": 1
       }
     ],
     "contracts": [
@@ -29,9 +32,9 @@
     ],
     "transactions": [
       {
-        "consensus_timestamp": 167654000123456,
+        "consensus_timestamp": 167654000123457,
         "payerAccountId": "5",
-        "transaction_hash": "519a008fabde4d",
+        "transaction_hash": "96ecf2e0cf1c8f7e2294ec731b2ad1aff95d9736f4ba15b5bbace1ad2766cc1c10554fd34ae23bf39931e9b90542b781",
         "valid_start_timestamp": 167654000123400
       }
     ]
