+++ hedera-mirror-rest/__tests__/specs/contracts/{id}/results/{id}/timestamp-transaction-not-found.json
@@ -19,22 +19,6 @@
         "timestamp_range": "[987654999123300,)"
       }
     ],
-    "contractresults": [
-      {
-        "amount": 30,
-        "bloom": [5, 5],
-        "call_result": [6, 6],
-        "consensus_timestamp": "167654000123456",
-        "contract_id": 5001,
-        "created_contract_ids": [7001],
-        "error_message": "",
-        "function_parameters": [7, 7],
-        "function_result": [8, 8],
-        "gas_limit": 987654,
-        "gas_used": 123,
-        "payer_account_id": 8001
-      }
-    ],
     "recordFiles": [],
     "transactions": []
   },
@@ -49,7 +33,7 @@
     "_status": {
       "messages": [
         {
-          "message": "No correlating transaction"
+          "message": "Not found"
         }
       ]
     }
